+++ src/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import BookForm from './components/BookForm';
-import BookList from './components/BookList';
+import BookForm from './containers/BookForm';
+import BookList from './containers/BookList';
 import './App.css';
 
 function App() {
