+++ src/components/Book.js
@@ -1,0 +1,15 @@
+import React from 'react'
+
+function Book(props) {
+    return (
+        <div>
+            <table>
+                <tr>
+                    {props.book}
+                </tr>
+            </table>
+        </div>
+    )
+}
+
+export default Book
