--- src/components/BookForm.js
@@ -1,11 +1,0 @@
-import React from 'react'
-
-function BookForm() {
-    return (
-        <div>
-            
-        </div>
-    )
-}
-
-export default BookForm
