--- src/components/BookList.js
@@ -1,27 +1,0 @@
-import React from 'react'
-
-function BookList() {
-    return (
-        <div>
-            <table>
-                <thead>
-                    <tr>
-                    <th>Id</th>
-                    <th>Title</th>
-                    <th>Category</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <tr>
-                        <td>id</td>
-                        <td>title</td>
-                        <td>Category</td>
-                    </tr>
-                </tbody>
-
-            </table>
-        </div>
-    )
-}
-
-export default BookList
